Rewrite observers.tests.test_state
Use parquet instead of CSV for data files because it can stored
fixed-length lists. Add utilties for creating CoordinateCovariances
which are full of nulls, since that's necessary when reading fixed
length lists with null data out of parquet.

Condense and parametrize the test case for observer state computation.

Obey lint